refactored text editor
