--- client/components/TextEditor.jsx
@@ -1,32 +1,0 @@
-import React from "react";
-import { EditorState } from "draft-js";
-import "draft-js/dist/Draft.css";
-import { Editor } from "react-draft-wysiwyg";
-import "react-draft-wysiwyg/dist/react-draft-wysiwyg.css";
-
-export default function MyEditor() {
-  const [editorState, setEditorState] = React.useState(() =>
-    EditorState.createEmpty()
-  );
-
-  const editor = React.useRef(null);
-  function focusEditor() {
-    editor.current.focus();
-  }
-
-  return (
-    <div
-      // style={{ border: "1px solid black", minHeight: "6em", cursor: "text" }}
-      // onClick={focusEditor}
-    >
-      <Editor
-        editorState={editorState}
-        toolbarClassName="toolbarClassName"
-        wrapperClassName="wrapperClassName"
-        editorClassName="editorClassName"
-        onEditorStateChange={setEditorState}
-      />
-    </div>
-  );
-}
-
